feat: extract package util
